# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Barcode Scanner Time Tracker
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      parameters:
        - name: activeOnly
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/barcode/{barcode}:
    get:
      operationId: getProductByBarcode
      tags: [products]
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /products/import:
    post:
      operationId: importProducts
      tags: [products]
      summary: Import products from CSV data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductImport"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"

  # Operations
...
  /reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: operatorId
          in: query
          schema:
            type: integer
        - name: shiftId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
  /reports/by-product:
    get:
      operationId: getReportByProduct
      tags: [reports]
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: operatorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Stats by product
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductStats"
  /reports/by-hour:
    get:
      operationId: getReportByHour
      tags: [reports]
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: operatorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Stats by hour
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HourStats"
  /reports/by-operator:
    get:
      operationId: getReportByOperator
      tags: [reports]
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Stats by operator
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OperatorStats"

  # Action log
...
  /export/operations:
    get:
      operationId: exportOperations
      tags: [reports]
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, json]
      responses:
        "200":
          description: Export data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportData"
  /export/products:
    get:
      operationId: exportProducts
      tags: [reports]
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, json]
      responses:
        "200":
          description: Export data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportData"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: integer
        barcode:
          type: string
        sku:
          type: ["string", "null"]
        name:
          type: string
        category:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        normTimeSeconds:
          type: ["integer", "null"]
        comment:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - barcode
        - name
        - active
        - createdAt
    ProductInput:
      type: object
      properties:
        barcode:
          type: string
        sku:
          type: string
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        normTimeSeconds:
          type: integer
        comment:
          type: string
        active:
          type: boolean
      required:
        - barcode
        - name
    ProductUpdate:
      type: object
      properties:
        barcode:
          type: string
        sku:
          type: string
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        normTimeSeconds:
          type: integer
        comment:
          type: string
        active:
          type: boolean
    ProductImport:
      type: object
      properties:
        rows:
          type: array
          items:
            type: object
            properties:
              barcode:
                type: string
              sku:
                type: string
              name:
                type: string
              category:
                type: string
              normTimeSeconds:
                type: integer
            required:
              - barcode
              - name
      required:
        - rows
    ImportResult:
      type: object
      properties:
        created:
          type: integer
        updated:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - created
        - updated
        - skipped
        - errors
    Operation:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: ["integer", "null"]
        operatorId:
          type: ["integer", "null"]
        shiftId:
          type: ["integer", "null"]
        workplaceId:
          type: ["integer", "null"]
        barcode:
          type: string
        productName:
          type: ["string", "null"]
        productSku:
          type: ["string", "null"]
        productCategory:
          type: ["string", "null"]
        operatorName:
          type: ["string", "null"]
        shiftName:
          type: ["string", "null"]
        workplaceName:
          type: ["string", "null"]
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, paused, stopped, completed]
        quantity:
          type: integer
        netDurationSeconds:
          type: ["integer", "null"]
        totalDurationSeconds:
          type: ["integer", "null"]
        pauseDurationSeconds:
          type: ["integer", "null"]
        normTimeSeconds:
          type: ["integer", "null"]
        comment:
          type: ["string", "null"]
        pauses:
          type: array
          items:
            $ref: "#/components/schemas/OperationPause"
        createdAt:
          type: string
      required:
        - id
        - barcode
        - startTime
        - status
        - quantity
        - createdAt
...
    OperationPause:
      type: object
      properties:
        id:
          type: integer
        operationId:
          type: integer
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
      required:
        - id
        - operationId
        - startTime
...
    ReportSummary:
      type: object
      properties:
        totalOperations:
          type: integer
        totalUnits:
          type: integer
        totalNetSeconds:
          type: integer
        totalPauseSeconds:
          type: integer
        avgOperationSeconds:
          type: integer
        avgSecondsPerUnit:
          type: integer
        aboveNormCount:
          type: integer
        belowNormCount:
          type: integer
        fastestProduct:
          type: ["string", "null"]
        slowestProduct:
          type: ["string", "null"]
        activeOperation:
          oneOf:
            - $ref: "#/components/schemas/Operation"
            - type: "null"
      required:
        - totalOperations
        - totalUnits
        - totalNetSeconds
        - totalPauseSeconds
        - avgOperationSeconds
        - avgSecondsPerUnit
        - aboveNormCount
        - belowNormCount
    ProductStats:
      type: object
      properties:
        productId:
          type: ["integer", "null"]
        barcode:
          type: string
        sku:
          type: ["string", "null"]
        name:
          type: string
        operationCount:
          type: integer
        totalUnits:
          type: integer
        totalNetSeconds:
          type: integer
        avgOperationSeconds:
          type: integer
        avgSecondsPerUnit:
          type: integer
        minSeconds:
          type: integer
        maxSeconds:
          type: integer
        normTimeSeconds:
          type: ["integer", "null"]
        deviationSeconds:
          type: ["integer", "null"]
      required:
        - barcode
        - name
        - operationCount
        - totalUnits
        - totalNetSeconds
        - avgOperationSeconds
        - avgSecondsPerUnit
        - minSeconds
        - maxSeconds
    HourStats:
      type: object
      properties:
        hour:
          type: integer
        operationCount:
          type: integer
        totalUnits:
          type: integer
        avgSeconds:
          type: integer
      required:
        - hour
        - operationCount
        - totalUnits
        - avgSeconds
    OperatorStats:
      type: object
      properties:
        operatorId:
          type: ["integer", "null"]
        operatorName:
          type: string
        operationCount:
          type: integer
        totalUnits:
          type: integer
        totalNetSeconds:
          type: integer
        avgOperationSeconds:
          type: integer
      required:
        - operatorName
        - operationCount
        - totalUnits
        - totalNetSeconds
        - avgOperationSeconds
...
    ExportData:
      type: object
      properties:
        filename:
          type: string
        mimeType:
          type: string
        data:
          type: string
      required:
        - filename
        - mimeType
        - data