# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Barcode Scanner Time Tracker
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  barcode: string;
  sku?: string;
  name: string;
  category?: string;
  unit?: string;
  normTimeSeconds?: number;
  comment?: string;
  active?: boolean;
}